Ignore .svn occurences in function makeSubstitutions().
git-svn-id: http://forge.voodooprojects.org/svn/chameleon/trunk@1782 b5af796a-00a8-11df-919b-ffff7a100b7d

+++ package/buildpkg.sh
@@ -166,9 +166,11 @@ $chameleonSubsts
 $ownSubst"
 
     for file in "$@";do
-        cp -pf "$file" "${file}.in"
-        sed "$allSubst" "${file}.in" > "${file}"
-        rm -f "${file}.in"
+        if [[ ! $file == *.svn* ]]; then
+            cp -pf "$file" "${file}.in"
+            sed "$allSubst" "${file}.in" > "${file}"
+            rm -f "${file}.in"
+        fi
     done
 }
